fix : 테스트 세분화 및 예약수정 테스트 추가
